Delay setting up uinput device + vdagent connection until we know the xorg res
Connecting to the vdagent virtio port will put the mouse in client mode,
which is not something we should do until the vdagent client has told
us the xorg resolution. So delay setting up the uinput device and
creating the vdagent virtio port connection until we know the xorg
resolution.