vdagentd: keep track of active session and xorg resolution
The uinput tablet device vdagentd creates MUST have the same resolution
as the active xorg session. So we remember to which session an agent belongs
and the resolution send by each agent and then when the session is changed
update the uinput device's resolution if necessary. Or close the vdagent
virtio channel if we don't know the resolution (fall back to server mouse
mode).