vdagentd: change socket name
Having a Unix Domain Socket directly under /tmp is a really bad idea
(think symlink attacks). The standard solution for this is to create a
subdir under /tmp. But since spice-vdagentd runs as root anyways we might
just as well put it in another much safer place. So now the socket is:
/var/run/spice-vdagentd/spice-vdagent-sock