vdagent-x11: Fixup NewValue Property Notify event filtering
When we received a NewValue Property Notify event for a property we
are not interested in, we would see this as an error, instead we should just
ignore the event.

In addition this patch also adds a check that the NewValue Property Notify
event is send from the window which we are actually interested in.