vdagentd: Close the virtio port when the active vdagent disconnects
We cannot use agent mouse mode without the session vdagent process, as we
need to know the resolution the client is running at. So we should
close the virtio port when the active vdagent disconnects, so that the
spice-server switches back to server mouse mode.

We were not properly updating active_session_conn when compiled without
consolekit support, which causes us to not close the virtio port, this patch
fixes this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>