randr: Round x monitor coordinates to a multiple of 8
We already do this when comparing new received monitor settings from the
client with the current ones, since the X-server rounds the width / x
coordinates to a multiple of 8 when we set them. Currently what happens is:
1) We get a monitor-config with a non multiple 8 width
2) We set the crtc to this width
3) We set the screen to this width
4) We configure the uinput tablet with this width
5) The X-server adjusts things to a multiple of 8
6) We re-configure the uinput tablet with the new width

By rounding the width and x coordinates before applying them we safe all
this back and forth to the X-server.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>