vdagent-x11: Fix crash on vdagend restart (rhbz#894365)
vdagent_x11_destroy() cleans up clipboard state and will tell the client
it is releasing the clipboard if it owns it. But when the vdagentd is
restarted the unix socket is disconnected and x11->vdagentd points to
already free-ed memory, leading to a crash.

This patch fixes this by not trying to send messages to the client on
cleanup when the unix socket is disconnected.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>