vdagentd-xorg-conf: Don't warn when there are too few qxl devices (rhbz#895004)
Since we now support more then 1 monitor / qxl device don't log a warning
when there are more monitors then qxl-devs. Note the warning does still
get written to the generated xorg.conf file, as that is intended for use with
xinerama setups, and there we do have a limit of 1 monitor / device.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>