vdagent: Add a cmdline option to select where to save file-xfers
The user can pass any arbitrary directory or the special values of
"xdg-desktop" or "xdg-download" to select the default xdg Desktop resp
xdg Download directories. If no value is specified the default is xdg-desktop
when running under a Desktop Environment which has icons on the desktop and
xdg-download on other Desktop Environments.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>