vdagent: Open file xfer save dir on file xfer completion
When not saving to the Desktop we need a way for the user to discover where
we've saved the file. This patch makes the session agent call xdg-open on
the file xfer save dir when a file xfer has completed in this case.

This is configurable from the cmdline by passing "-o 0" or "-o 1"

Signed-off-by: Hans de Goede <hdegoede@redhat.com>