console-kit: implement check for locked session
We register to read the Lock, Unlock and IdleHintChanged signals from
ConsoleKit.Session. The Lock/Unlock signals should be the right
signals for the job but not all Desktop Managers implement its locking
in a way that trigger those signals. That's why we double-check with
IdleHintChanged signal that it might be triggered by other services
like screen savers.

Acked-by: Jonathon Jongsma <jjongsma@redhat.com>